realized it makes more sense to keep these details here lol
